url수정
url수정과 view수정
