+++ .gitignore
@@ -174,3 +174,4 @@ cython_debug/
 # End of https://www.toptal.com/developers/gitignore/api/django
 
 /venv
+/myvenv
