+++ static/js/currentPos.js
